+++ CHANGELOG.md
@@ -23,7 +23,7 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
-*   Add `tf.nest` compatibility for `DataclassArray`.
+*   Add `tf.nest`/`tf.data` compatibility for `DataclassArray`.
 
 ## [1.4.2] - 2023-07-10